make freeverb #![no_std] compatible and use f32

+++ src/freeverb/src/all_pass.rs
@@ -11,7 +11,7 @@ impl AllPass {
         }
     }
 
-    pub fn tick(&mut self, input: f64) -> f64 {
+    pub fn tick(&mut self, input: f32) -> f32 {
         let delayed = self.delay_line.read();
         let output = -input + delayed;
